moves model tests to one module per TestCase class
